Scale the pose result of 2D-2D matching, plus misc fixes
Now the results are more or less passable. Of course, without having a
calibrated baseline, we don't know the true scale of the pose decomposition,
so this cheats a bit by injecting the scale of the ground-truth poses. If we
had 3D points or a depth map, we could fix the new pose to the scale of the
ground truth DB.